requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.4
django-admin==1.3.2
django-excel-response2==2.0.8
django-heroku==0.3.1
django-import-export==1.2.0
django-six==1.0.4
et-xmlfile==1.0.1
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
psycopg2==2.8.3
python-dateutil==1.5
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
screen==1.0.1
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
